+++ android/SSPSDK.kt
@@ -821,4 +821,4 @@ sealed class SDKError : Exception() {
 }
 
 // SDK version
-const val SDK_VERSION = "0.5.16"
+const val SDK_VERSION = "0.5.17"